TUGAS 5

+++ app/Http/Controllers/ProductController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Product;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ProductController extends Controller
 {
@@ -61,17 +62,22 @@ class ProductController extends Controller
 
     public function edit(Product $product)
     {
+        Gate::authorize('update', $product);
+    
         $users = User::orderBy('name')->get();
-
         return view('product.edit', compact('product', 'users'));
     }
 
     public function delete($id)
     {
         $product = Product::findOrFail($id);
-
+        Gate::authorize('delete', $product);
+    
         $product->delete();
-
-        return redirect()->route('product.index')->with('success', 'Product berhasil dihapus');
+        return redirect()->route('product.index')->with('success', 'Product dihapus');
+    }
+    public function export()
+    {
+    return "Halaman Export Data (Hanya Admin yang bisa lihat ini)";
     }
 }
